[MovieList, MovieSelection] Add optional duration column to list
[MovieSelection]

Add config variable to control the visibility of the duration/length
column. Use a ConfigSelection rather than ConfigYesNo (or similar
boolean config) so that the choice can be extended when issue #499
is implemented.

[MovieList]

Add code to optionally display a duration column. Only gather the
data needed when the column is enabled.

Change the call of moviePlayState() to pass a length of 0 instead
of data.len. This preserves the former behaviour. In any case,
data.len is an inappropriate value, since it is time in seconds,
while moviePlayState() expects the recording file length.

Use expression if to calculate the duration text instead of an if
statement for compactness.

Rearrange the calculation of the row field positions to make the
position clearer and to make it easier to rearrange the colums in
the code.