[MovieBrowserConfiguration] Correctly hide/display inactive options
Restructure the creation of the option list so that it can be called
after the screen has been created.

Change the test on config.usage.show_icons_in_movielist so that it
compares to the "off" value of the ConfigSelection rather than
assuming that the config item returns a boolean-like value.

Use Notifiers to rebuild the option list whenever
config.usage.show_icons_in_movielist changes.

Make more use of temporary lists to group the addition of
"always-displayed" settings.