[font.ccp] This fixes a BSoD on viewing logs of a certain size.
It is clear what the problem was, or rather where the problem occurs, and this fix can stop it happening. What is not clear is why the problem occurs without this fix. So perhaps it's more of a workaround.

lib/gdi/font.cpp contains this code:
Code:

                // gaaanz lahm, aber anders geht das leider nicht, sorry.
        FriBidiChar array[size], target[size];

This dynamically assigns those two arrays in the stack (FriBidiChar is a wchar_t - 4 bytes).
This all works OK. For a crashing log the size was 262691, and the arrays were assigned on the stack at address which were 1050768 bytes apart (2*262691 is 1050764, so there's some byte-alignment for efficiency).
The fribidi_log2vis function in libfridi copies array to target using memcpy(). This produces a SIGSEGV at the first 16384 page-aligned reference (so 0x0000z000, z == 0, 4, 8, c) into the target array. It does the same if memmove is being used. This also happens if I use alloca() to "manually" assign the arrays' space on the stack. If an assignment loop is used it gets the SEGV 4 bytes earlier (which is, probably, just a hardware reporting feature).
Increasing the stack size (from 8MB to 16, then 64) didn't seem to make any difference either beyond moving the addresses.

However, allocating those arrays in the data segment using malloc() (and remembering to free them!) does work - at least for the ones that have been tried so far...