Reverted again as there is too much opposition. We need a coder to create a multi contents template instaed of the percentage based template we have now.
Reverted, as this is required by the new EPG-search options. The lay-out is fontsize/width dependant (a percentage of), so the actual lay-out in the skin is very much dependant on the font size used and the allowed width. This commit has been thouroughly tested, using both orbital position On and Off, using the skins ViX-HD, Magic-HD, Magic-FHD & YouVix; other skins might need to be adapted to these values.

Revert "[EpgList] Tweaked lay-out (again)."

This reverts commit c58fd84c2645362da54200055c5ac8856b32826b. (reverted from commit cee467569bc3b2d94c91bbb1a6a1ea2b71166575) (reverted from commit e83580d309f5bfd97443ffc808434a49364aa408)