Fix for zaptimer not going to requested service when executed when movieplayer is active. Signed-off by Burdman
A Zap timer is a record timer which only changes the viewed channel.
Say you device to set on to switch to BBC1 at 9pm.
At 8pm you are actually watching ITV1.
Then you start to watch a previously recorded film.
9pm arrives, the Zap timer cuts in.
You should now end up on BBC1 (the target of the Zap) but you actually end up on ITV1 (the channel that was being viewed before you started to watch a recording).
The reason is that the current code does the Zap first, but then the MoviePlayer shuts down and that has a side effect of putting you back to the channel it was on before the playback started (see #4 in that thread), hence "undoing" the Zap.
This fix ensures that the MoviePlayer is shutdown before the Zap takes affect.