[ChannelSelection] This fixws the problem that E2 doesn't always knows in which bouquet it is. Signed-off by prl001.
In Screens/ChannelSelection.py, there's the quite reasonable-looking code:

	class ChannelSelection(ChannelSelectionBase, ChannelSelectionEdit, ChannelSelectionEPG, SelectionEventInfo):
		instance = None

		def __init__(self, session):
			...
			assert ChannelSelection.instance is None, "class InfoBar is a singleton class and just one instance of this class is allowed!"
			ChannelSelection.instance = self

The problem is then that there's also:

	class PiPZapSelection(ChannelSelection):
		def __init__(self, session):
			ChannelSelection.__init__(self, session)

and:

	class InfoBarChannelSelection:
		""" ChannelSelection - handles the channelSelection dialog and the initial
		channelChange actions which open the channelSelection dialog """
		def __init__(self):
			# instantiate forever
			self.servicelist = self.session.instantiateDialog(ChannelSelection)
			self.servicelist2 = self.session.instantiateDialog(PiPZapSelection)

This means that `ChannelSelection.instance` is initially (correctly) set to be the instance that's saved in `InfoBarChannelSelection.servicelist`, but then it's over-written to be the PiPZapSelection instance that's assigned to `InfoBarChannelSelection.servicelist2`.

That means that in all places in the code where `ChannelSelection.instance` is referenced (like in EPGSearch), `ChannelSelection.instance` will refer to the PIP channel changer instance, not the normal channel changer instance.

Superficially, you'd think that the assert would stop this happening, but the python modules are normally compiled with optimisation on, which disables any asserts. Also, at least on the Beyonwiz image, python itself is compiled in a way that `__debug__` is always `False`, so assertions would never be invoked, even if they were compiled into the code.