[EpgList] Adjusted lay-out. Signed-off by prl001.
New column width code:

fontSize = self.eventFontSizeSingle + config.epgselection.enhanced_eventfs.value
self.weekday_rect = Rect(0, 0, int(fontSize * 2.3), height)

The fontSize is the nominal size of the font, which is derived from the sum of the font size set in the code (28 for UI overlays with a pixel width of 1980 and 22 otherwise) and the "Event font size" in the settings for the "Show EPG for current channel" EPG.

That makes the column width 64 if the UI overlay horizontal resolution is 1980, and 50 otherwise if the "Event font size" is at its default of 0.

The "Event font size" setting can't be changed in the EPGSearch MENU>Setup settings.

The weekday column width is set as int(fontSize * 2.3), i.e. 2.3 times the nominal size of the font, rounded down to the nearest pixel.

If you want to change the width of the weekday column, change the 2.3 in the assignment to self.weekday_rect. If you want to change the width of the date/time column, change the 5.8 in the assignment to self.datetime_rect.

The other change is to make the description rectangle self.descr_rect use the whole of the rest of the column width:

self.descr_rect = Rect(self.datetime_rect.left() + self.datetime_rect.width(), 0, width - self.datetime_rect.left() - self.datetime_rect.width(), height)

I.e. the width of the description box is the width of the list, less the x-coordinate of the left edge of the description box.

The original VIX code made the description rectangle 62% of the list width. With the other two boxes being 10 and 25% respectively, that meant that only 97% of the width was used.