service db: add support fro writing a new lamdeb format
This patch implements a new lamedb format, version 5.
It is easier to read for humans and non-humans.
Inforation is now line based. lines starting with a 't'
define transponders, lines starting with an 's' will define a service.
Other lines are ignored but it is considered good practice to prefix
other lines with a '#' character to predicate it as comment.

The version 5 file will start with a set of comment lines that
describe the format. I.e. the files is self descriptive.
Note that some (service) parameters are optional. This is
indicated in the descritpion by surrounding the parameter with []
(those [] are not added in the actual data off coarse).

The current code writes a version 4 (lamedb) and a version 5
file (lamedb5). The parser verifies the version number and uses
the correct parsing code, but as it still only reads lamedb it
will by default run the version 4 parser.

At a later state the code for writing version 4 files could be removed
and the file reader reads the version 5 file unless a version 4
file is available that has a more recent date. Or the version 5
data is just written as 'lamnedb' and the reader just decides on
the version number in lamedb.
At an even later state, when all settingedit tools have adopted this
new format, the version 4 code could be removed all together.

We will see what the future brings...