Optimize buffersize for streaming. Signed-off by enf70.
The streaming buffer was hard-coded in the demux.h file to 4 buffers (of 188KB) which for a HD channel accounts for less or around 300ms of buffer time. For wired streaming this is fine (most of the times) but for higher latency (wireless for instance) streaming this buffer is too small and causes intermittent freezes and stutters in the streaming (more noticeable when streaming HD channels).
This patch makes the streaming buffer use the same value as the recording buffer which is calculated in the demux.cpp file depending on the box RAM size and it starts at a more decent value of 8 buffers (of 188KB) for the smallest box.