Fix Bug #513: Incorrect BLUE button text in Location selector screen
[LocationBox]

Initialise the BLUE buton help text from the onLayoutFinish callouts
instead of from onExecBegin, by moving the button text initialisation
code into switchToFileListOnStart().

The setting the button text from onLayoutFinish is late enough to
avoid problems with the double widget definition of the button, and
avoids having the text reset to "Add bookmark" called each time
LocationBox was returned to from a popup, as was the case when
called from onExecBegin.