Fix bug #525: [Channel Selection] "Now" instant recording not highlighted on creation
[ServiceList]
Add method updateCurrent() to update the display of the current entry
in the setvice list. This happens as a side-effect of disabling then re-enabling
selection on the list instance in the method.

[ChannelSelectionEPG]
Use ServiceList.updateCurrent() to update the state of the
current selection in the service list after a "now" timer has
been added or deleted.