Fix Bug #515: IceTV-managed timers deleted twice
[EPGFetcher]

In onTimerChanged(), use entry.cancelled to determine wthether the
method has been called because the timer has only been changed
(entry.cancelled is False), or whether the method has been called
in a step in removing the timer (entry.cancelled is True), in which
case onTimerRemoved() will be called later and the timer can be
processed once there.

So if entry.cancelled is True, onTimerChanged() returns immediately.