Fix Movielist date width reversion
[MovieList]

Make the default width of the date/time field proportional to the
font size, instead of fixed 150px. Defaults to 9 * font size (reduced
from 10 * font size).

Make the font size multiplier skinnable as attribute dateWidthScale
(interpreted as float) in the movielist's skin widget.

If the dateWidth attribute is set in the movielist's skin widget
it overrides the width from font size calculation and a fixed width
is used instead.

Apply the date column width to Directory and Trash entries.