[EPGList] Fix reversion of weekday column width
Increase the width allocated for the "weekday" column for the wider
content placed inthe column in the Beyonwiz code for the current
channel EPG and the similar events lists than is used by in the
OpenVIX code for the "weekday" column.

This makes the corresponding column in the EPG Search screen too
wide for its content.