Fix bug #530: "Stop service on return to movie list" setting does nothing
Remove code and config list entries associated with
config.movielist.stop_service.

Leave config.movielist.stop_service in place for backward compatibility,
but flag it as unused. Add code to set it back to its default value
if it has a non-default value to clear it from /etc/enigma2/settings.

Move the code used by config.movielist.stop_service to restart
timeshift recording at the end of playback if live TV starts in the
background of the selection list.