Fix issue partial #537: Reload "resumepoints.pkl" only when needed
[InfoBarGenerics]

Use Tools.Directories.resolveFilename() to set the path to
"resumepoints.pkl" instead of having the path hardwired in the code.

Only reload the resume points file in updateResumePointCache() if
it has been modified outside the cache system.

Allow getResumePoint() to return a resume point for DVB recordings
and other ".ts" files (serviceref.type == 1).

Regroup the globals used by the cache so that they are mostly
together at the start of the cache code.

Also remove unused "global" declarations in the cache functions.