Fix issue #537: Remove resume point cache entries when directories are deleted
This commit completes the enhancement implementation.

[InfoBarGenerics]

Add code to InfoBarGenerics.delResumePoint() delete all the resume
point cache entry entries contained in the serviceref argument if
the serviceref has the eServiceReference.mustDescent flag set.

[MovieSelection]

Use the new functionality of InfoBarGenerics.delResumePoint() to
delete cache entries for media and recordings when a folder is
deleted or moved to trash, and when either a recording/media file
or a directory is moved or renamed in the media player.