cablescan: limit the SDT timeout to 5s
when channels are not available, we are waiting for the SDT timeout (60
seconds) before we skip to the next channel.
This happens because we do not connect a channel state event handler, so
we do not get notified of tune failures.

Instead of using the channel state, we can limit the SDT timeout, so we
wait for a maximum of 5s per channel.
When no SDT is received for 5s, we cannot use the channel anyway.