[ePicLoad] Fix fill reversion - incorrect 8-bit fill
This fixes a problem seen in the 16.1 alpha, where block of incorrect
fill was displayed along the right-hand edge of the remote control
image widget.

The bug that caused this problem was in ePicLoad::getData(), and
was caused by the fill code using 32-bit fill on 8-bit images.

Made fill correct for both 32-bit and 8-bit fill.

Moved the 32-bit fill into a local function to hold otherwise
duplicated code: fillrow_uint(). Used memset() for 8-bit fill.