[ePicLoad] Fix fill reversion - incorrect edge fill
This fixes a problem seen in the 16.1 alpha, where if the scaled
image is 1 pixel smaller than the widget dimensions in either
direction.

This is because xoff/yoff are rounded down when calculated, meaning
that in this case, xoff (resp yoff) are zero if the 1-pixel size
difference is on that axis.

This fix ensures that the right/bottom 1-pixel column/row is filled
when the difference on that axis is 1.

There is a simpler fix, by changing:

xoff = (m_filepara->max_x - scrx) / 2

to

xoff = (m_filepara->max_x - scrx + 1) / 2

also respectively for yoff.

However, that changes the alignment of the pixmap within the fill
slightly, so the simpler fix wasn't used.