[TimerList] Tidy up getOrbitalPos()
Add code to TimerList.getOrbitalPos() to display HDMI IN for HDMI
IN services instead of an orbit location.

In the process tidy up the field extraction code in the method so
that all the field operations can be done using a single split.
Make the Stream identification a bit more explicit (restrict it to
the NAME field of the serviceref.

This isn't strictly part of the reversion fix.