[TimerList] Fix orbpos reversion
Don't display the orbpos text on Beyonwiz firmware. Use the recovered
space for the repeat type text.

Move the default position for the orbpos text slightly to the left
to gain a little more space.

Rename "Everyday" repeats to "Daily" because "Everyday" is wrong
in this context, and "Daily" matches the TimerEdit screen value.