Fix bug #544: Software Management plugin settings can leave Software Management inaccessible
Change the interaction between the Software Management plugin and GeneralMenu
so that:

The Software Manager is always shown in the Plugin Browser (PLUGIN
from live TV).

The Software Manager is never shown in the Plugins section of the
main menu (MENU in live TV). There are only two spare slots in it
on the T4 before it starts to need to scroll. This is unchanged
from the current code.

The "Show software manager in plugin menu" option is changed to
read "Show software manager in tasks/setup menu", the test in
SoftwareManager on displaying it is changed to "id_mainmenu_tasks"
so that it inserts itself in the Setup/Tasks submenu of the main
menu and the exclusion in the main menu that blocks the Software
Manager from appearing in the non-Plugins submenus is removed. I
think this is in keeping with the plugin code's intention, to display
itself in the main menu submenu with menuID "setup".

The currently working option to have the Software Manager insert
itself in the BLUE button Extensions menu is retained.

Allow for information text on all entries in the list, not just for
"Overwrite configuration files?"

Add information text to the config list entries for "Overwrite
configuration files?", "Show software manager in tasks/setup menu"
and "Show software manager on blue button".