atsc: add EIT support
providing now/next information

Currently, only EIT-0 is parsed.
The matching ETT-0 is used to provide event descriptions.
Both titles and descriptions are selected based on the preferred EPG
languages.

For both EIT and ETT, only plain messages with UTF-16 encoding are
parsed.
Huffman and SCSU are not yet supported by the current multi-string-
object implementation.

Side-effect:
eDVBServiceEITHandler contained a ePtr<iDVBDemux> member, which was
never used. (it should have been used to make sure an additional ref was
held on the demux, while the EIT reader was running)
The current implementation uses the member, thus adding refs on the
demux.
If resource refcount issues are starting to occur, this might need a
closer look, there might be an issue explaining why the ePtr<iDVBDemux>
member was never assigned.