Update ePicLoad::setPara() backgrounds to use 0xFF alpha
Where appropriate change, the background colour alphass in calls
of ePicLoad::setPara() to fully transparent 0xFF to compensate for
0xFF no longer being forced as the alpha in ePicLoad::setPara(),
fixed in the fix for bug #542.

Because of bug #545, this means that the fill for non-colourmapped
images is for now opaque when it should be tranparent and vice
versa.

[IniSHOUTcast]

Change the event filtered out of debug logging from 17 to 18.  Event
17 is evVideoProgressiveChanged (which doesn't seem appropriate at
all) and 18 is evBuffering. This stops a lot of debug file clutter.
Presumably at some time the numbering changed. These numerical
event codes should be replaced by the correct symbolic codes from
iPlayableService_ENUMS.