[PluginList.py] Correct code error and improve defaults
The text horizontal location offset on line 21 should be "dx" not "nx".

A default width of 1280 is a more appropriate limit for a HD screen.  The current nominal default of 440 does not allow for user interfaces that define the UI element to be wider than 440 pixels.  The 440 pixel limit could cause the text to be inappropriately truncated.  The limit of 1280 was chosen as this is the largest possible width for generic HD UIs.