[PluginList.py & skin.py] Add a new parameter and font to allow improved category display
[PluginList.py] Add a new parameter and font to allow improved category display

[skin.py] Add new FHD parameter for the new plugin category display

This change allows the plugin list categories to be displayed using a different parameter and font rather than being forced to share the plugin name definition.  The plugin name is constrained to approximately half the height of the cell typically aligned to the top of the cell..  The new font and parameter do not have this constraint and hence allow a larger and more centred version of the category names.  This matches the category name display before the introduction of the parameters system.