[ServiceScan.py] Remove unnecessary progress limit of 99%
The underlying code in the progress counter getProgress() can only return numbers between 0 and 100.  (Confirmed by Prl.)  There is no reasonable reason to limit the displayed output only up to 99.  If there was a perceived issue that the number could go above 100 then the protection should be so that the number can not exceed 100.

In the user interface users get to see the scan progress starting at 0% and increasing.  Due to this code the number can never reach 100%.  To many users this is an indication that the scan may not have completed correctly.

As there appears to be no technical justification for artificially limiting the percentage counter to 99% and to allay user concern that scans are not properly completing the display limit is being removed.