[PluginList.py] Revert default width change
I still maintain that a width of 440 pixels is a random choice and that there is logic to change the fall-back default to 1280.

Having checked the PluginBrowser screen that calls this module I note that all calls do provide an appropriate width that relates to the current skin control width configuration.

To aid in future merge efforts I will revert this change.