Fix bug #534: Enabling parental control prevents access to .xxx directories
[MovieSelection]
Add a default parameter to gotFilename() to indicate that a correct
access PIN has been entered to avoid being stuck in a loop where
the PIN is repeatedly requested even after it has been entered
correctly.

Narrow the test on allowing non-PIN-controlled access to Trash
directories from

baseName.startswith(".Trash")

to

baseName not in (".Trash", ".trash", ".Trashcan")

to cover cases discussed in the issue comments.