Restore EPG display of highlights on all current programs
This commit restores previous functionality where the various EPG views displayed a highlight for all events that are currently in progress.

This incarnation adds a new configuration setting "config.epgselection.graph_shownow" that controls this display option.  For compatibility with recent firmware expectations this setting is defaulted to "Off" (False) so that the highlights are not displayed.