3D ViX Resolution impacts LCD initialisation and the ET8500 LCD skin.
The 2015 3D change for OpenViX modified 5 modules and in particular used UserInterfacePositioner.py (lib/python/Screens) to update the 3D config files. The change needed the call in mytest.py for UserInterfacePositioner to be moved to an earlier position to get the implementation to work or it resulted in a failure in Avswitch; fine except that this meant that the LCD was initialized with the OSD settings - causing LCD skin issues for receivers such as the ET8500.

This commit basically reverts the situation so the 3D and OSD initialization can be done in 2 mytest.py calls at the correct time for both - so OSD functionally works as it did before the 3D change.
Credits TWOL.