e2avahi: Use hostname as service name
It's rather unclear what service name is supposed to be, but other
services like NFS and Samba use the host name there. Apparently the
name+type must be unique on the network, or you'll get a collision.
Implement this by defaulting to the local hostname when the name
is blank or NULL.

Advertise the streamserver as such.