[MovieSelection] Make adaptive/fuzzy dates optional
This change adds a new setting to the MovieSelection settings menu called "Use adaptive date display".  This option defaults to "Yes" which was the previous behaviour.  When set to "No" all dates in the MovieSelection list are shown as day, date and time with the date including the year..  The abbreviations like "Today" and "Yesterday" etc are not used.