A new solution regards fastscan where <n/a> appears in channel list
As the previous solution has a nasty side effect. Now in the
service_orbital_position "dictionary" I added the getOriginalNetworkId to
make the ServiceID more unique. It now fix the <n/a> issue without the
additional loops and it does not have the side-effect as mentioned here: https://forums.openpli.org/topic/47379-fastscan-tv-vlaanderen/page-2

Thanks fenzo for reporting!!!