Add compact date format (#101)
* Add compact date format

Add a new format "config.usage.date.compact" and a matching ClockToText converter format "CompactVFD" to compress the date display to a format usable by smaller VFD displays.

This is accomplished by removing spaces and hyphens from the display.  To maintain clarity of the information dates with the month as a word are always displayed as DayMonTime.  That is, "14Jan12:34PM".  Numeric months will retain the "/" separator and the location appropriate order.

* Correction: Separate numeric data from time

Add a space to separate day of month or numeric month from the time.