Removed /etc/enigma2/automounts.xml from backup process
Currently the automounts.xml gets backup'd 2 times. (can be checked in console output and opening the backup file manually)
From what i can see this happend becaurse the complete /etc/enigma2/ folder is backup'd. And then again the induvidual /etc/enigma2/automounts.xml is backuped.
So this last enrty can be removed, as the file is already backup'd with the complete folder.
But please review.