Fix bug #554: Channel change from EPG called from MoviePlayer plays live TV in the MoviePlayer
[EPGSelection]

In EPGSelection.zap() use 'close' as the parameter to self.close()
instead of True. This closes the MoviePlayer and returns control
to InfoBar, as in EPGSelection.zapTo().

This means that when long-OK (Zap+Exit) is used in the EPG from
Movie player, bug #554 is fixed, but now both the short-OK (Zap)
and long-OK (Zap+Exit) paths out of MoviePlayer via the EPG suffer
from bug #553.