[Time] Improve Date / Time front panel support (#123)
* Add compact date format

Add a new format "config.usage.date.compact" and a matching ClockToText converter format "CompactVFD" to compress the date display to a format usable by smaller VFD displays.

This is accomplished by removing spaces and hyphens from the display.  To maintain clarity of the information dates with the month as a word are always displayed as DayMonTime.  That is, "14Jan12:34PM".  Numeric months will retain the "/" separator and the location appropriate order.

* Correction: Separate numeric data from time

Add a space to separate day of month or numeric month from the time.

* Improve Date / Time front panel support

- Introduce two new config elements "config.usage.time.display" and "config.usage.time.wide_display" and an associated ClockToText converter option called "Display".  This is identical to the capabilities of "Default" ClockToText option but it is controlled by a separate time setting UI in the Front Panel Settings screen.

  It is the "config.usage.time.display" time format that will be used in all VFD displays. The associated "config.usage.time.wide_display" controls the new dynamic formatting of the "config.usage.date.compact" as described next.

  The new "Display" ClockToText option can be used in all LCD and OLED front panel display skins and will keep the UI time formatting separate and independent from the front panel time formatting.

- Change the meaning of the config.usage.date.compact item. This is the date format used by the 12 character VFD ClockToDate options (VFD and VFD12). This was traditionally a statically defined item. The revised code now modifies the format of the date based on the selected date format and then adjusted to suit the available space remaining based on the time format selected.

In other words the day and month are separated by the selected separator and the date and time are separated by a space if a 24 hour clock is selected. The day and month as well as the date and time separations are removed if a 12 hour clock is selected (the current mode of operation). This ensures that the date and time string is never longer than 12 characters but tries to preserve spacing, for better visual appeal, if this can be done without exceeding the available space.

- A new parameter "AllowUserDatesAndTimesDisplay" has been created for use in display skins to indicate that the skin supports the new display time formats.  The parameter syntax is:
    <parameters>
      <parameter name="AllowUserDatesAndTimesDisplay" value="1" />
    </parameters>

  Use the syntax as described to enable the use and selection of the new display skin time formats.  A missing parameter or a value of "0" will disable the feature and force all display skin times to the standard 24 hour clock format.