[EpgList.py] Add EPGSearch column match flag
EPGSearch uses the column definitions defined for the EpgList SingleEPG.

This commit adds a flag that is used by the EPGSearch plugin to match its displays to the Beyonwiz SingleEPG format.  That is, the flag will cause EPGSearch to display the day and date in column one and the start and end times in column two.