Make USB update recommendation more prominent.
Some users ignore the recommendation to use the USB update method and
claim that there was no notice shown, that it was not prominent enough
or that it was shown at a time where they were not expecting it.

Address those complaints by making the notice more prominent and show
it in multiple places.