[EpgList.py] Update MultiEPG to use enhanced new date/time formats
This variation on the previous change splits the times column into two so as to maintain the right alignment of both the start and end times.

The duration time column was previously hard coded as an extract of the event column.  It is now its own column and can be size adjusted in the skin like the other columns.

This change also standardises the EPG font size for a more consistent presentation.  The font size difference between the service and the times / event has been removed for a more balanced display.