[skin_display] Adjust for enhanced time formats
[SummaryClockPanel]

Increase width to allow for wider AM/PM & am/pm times.

Reduce font size to match the text that's usually to the left of
the clock and align with that text's vertical position and use the
same valign as that text. This means that the additional width
needed is not as great as would have been with the previous font
size.

[SummaryTunerStatusPanel]

Move the tuner icons left to accommodate the increased witdh of
SummaryClockPanel.

[TimerEditListSummary]

Reduce width of the duration and state fields to allow for the extra
width of the clock.

[StandbySummary]

Change the clock format from the fixed 24-hour HH:MM to the
user-controllable Default.

Decrease the font size so that the same vertical space can include
the descender on "pm". The time field vertical size and position
was also changed to not overlap the date field vertically and to
be within the screen area. The width of the time field was increased
to accept the AM/PM  and am/pm styles.