Fix bug #558: Odd scrolling in RollerCharLCD and RollerCharLCDLong
Change RollerCharLCD to use the actual pixel widths of the text,
the space character and the widget to calculate the amount of padding
in the marquee text.

Always display the whole remaining marquee text, instead of a limited
amount of it, so that it always uses the full widget width.

When the scrolling is finished set the text to the original text.

Remove dependence of the code on model types.

This fix makes the string length workaround in RollerCharLCDLong
unnecessary, so RollerCharLCDLong now simply inherits from
RollerCharLCD and sets a different scroll start delay, which is its
only remaining difference. It's mostly retained for backward compatibility.

Because eLabel::calculateSize() can't return a width greater than
eLabel::size(), this code will scroll if the text width is exactly
the same as the widget width. I don't think that can be avoided.