[UsageConfig.py] Remove prototype code (#132)
Now that the new date / time code is out in the wild the management user interface appears to be working as designed and users appear to accept the format selection methodology.

I now feel that the prototype alternatives for the UI selection interface can now be removed.  This significantly reduces the size and complexity of the code without any change in functionality or user interface.