Expand date / time formats (#133)
This change updates the config screen prompt strings for the date and time settings.  Rather than using the generic "9" marker the displays use more descriptive tokens.  "D" and "DD" for day of the month, "M" and "MM" for the numeric representation of the month, "H" and "HH" for the hour in a 24 hour clock, "h" and "hh" for the hour in a 12 hour clock, "mm" for the minutes of the time and "ss" for the seconds of the time.

With these new distinctions in place some more date and time formats have been added to both the date and time options.  With the previous prompt labels some of these options were not viable.

Add a new ClockToText option called VFD11 that addresses issues on the Beyonwiz T3 that only has an 11 character VFD display.  To add this option the processing logic needed to be slightly adjusted.  These changes should have no effect on any other hardware and only take effect when the Beyonwiz T3 display skin uses the VFD11 rather than the VFD12 option.