Fix bug 553: Timeshift buffer doesn't start after channel change from EPG called from MoviePlayer
When InfoBarEPG.reopen() is a method in MoviePlayer and is called
with parameter 'close', indicating that a service has been zapped
to in EPGSelection, save the running service as self.lastservice
(as in previous code) and then stop play.

This means that when self.lastservice is played in MoviePlayer.__onClose(),
self.lastservice will be restarted in the correct context so that
its evStart event is directed to the InfoBar.