Fix bug #561: Channel preview in EPG called from MoviePlayer plays live TV in the MoviePlayer
[EPGSelection]

In closeScreen(), play the media or recording again if EPGSelection
was started with playback in progress and Channel preview mode is
enabled for the EPG style. Play was already started for media or
recordings if Channel preview mode is disabled.

In both cases use new InfoBarCueSheetSupport.forceNextResume()
method to force resuming at the point where playback was stopped
for channel preview.

[InfoBarCueSheetSupport]

Add the ability to override the user setting of "Behaviour when a
movie is started" for the next playback start, so that uses use the
resume point unconditionally (if it's valid).

New method forceNextResume() and supporting code in ____serviceStarted().