Fix Bug #562: Timeshift control malfunctions after exit from MoviePlayer
Remove the code that tries to start timeshift when live TV playback
resumes in MovieSelectionScreen. Starting timeshift this way appears
to be the cause of the later problems when timeshift is used on
return to live TV.

A new enhancement request needs to be created to allow timeshift to start
while still in MovieSelectionScreen.