[NimManager/ScanSetup/Satfinder/etc] Skip multistream transponders.
When getting a transponder list from nimmanager.getTransponders(), only retieve multistream transponders when tuner is multistream capable. Otherwise just return normal transponders. If frontend id is not supplied, the complete list will be return as previously. This allows compatiblity for plugins that are shared with other teams (e.g. Blindscan) that do not have this feature in their code.