Fix Bugs #569 #570 #571: Various related no-preview mode EPG-from-MoviePlayer bugs
The fixes for these buga can't be readily separated, there are too
many interdependencies in the fixes.

[InfoBarEPG]

In openMultiServiceEPG() set up self.StartBouquet and self.StartRef
in the same way as in openGraphEPG(). There doesn't appear to be
any reason for openMultiServiceEPG()'s settings of those attributes
being different from the settings in openMultiServiceEPG().

This fix means that in the comparison between
Navigation.getCurrentlyPlayingServiceOrGroup() and EPGSelection.StartRef,
the correct value for StartRef is used. It's a different self, but
EPGSelection.StartRef is initialised from InfoBarEPG.StartRef.

[EPGSelection]

In closeScreen(), simply keep playing the same service on EPG exit
if the EPG preview mode is disabled.

If Preview is not enabled for the EPG, set the parameter on the call of
close() to 'close', so that if the EPG has been opened from inside
the MoviePlayer, the movie player is exited and control returns to
the InfoBar.