[NimManager] Fix issues with multi type tuners
Using DVB API to create the multi type list results in the dictionary keys
not matching the values actually used in /proc based mode lookup table.

This results in strange issues, such as a DVB-T2 tuner being switched
from DVB-T2 mode to DVB-C mode, or even worse, tuner modes being switched
repeatedly.

To improve predictability, default a multi type tuner to DVB-T2 mode if
available, otherwise default to DVB-T mode and if that isn't available
either, do not provide a default and fall back to the first mode in the
list.