Fix Issue #577: [AboutBase] Remove use of native key bindings
Requires the fix of Bug #576 (commit 2e05b54) for correct operation
of paging.

[AboutBase]

Remove setBindings() & restoreBindings() and the call to setBindings()
to remove dependence on native key bindings.

Remove no longer used imports of enigma.eActionMap and enigma.eListbox.

Increase the priority of the self["actions"] ActionMap so that the
"up" and "down" actions in the "DirectionActions" over-ride the
native bindings for UP & DOWN.

Add "up" and "down" actions to self["actions"] and bind them to
pageUp() and pageDown(), which after the fix of Bug #576 will page
correctly.

Add pageUp() and pageDown() methods to implement paging via
self["actions"].