CustListEditor: additional options; changed keys.
Change long YELLOW/BLUE to use the minute seek.  Add BACK to restore
the initial cuts, or remove all cuts.

Add "enable cuts" to the menu to preview the cuts; it then changes to
"disable cuts" to go back to free browsing.

Add "execute cuts" to the menu to run the "Execute cuts" plugin, which
will also close the editor.

Just use the short press text for RED & GREEN, to fit the standard
button images.