Show directory & file counts for movie directories
Instead of displaying "Directory" for a directory in the movie list,
optionally show the count of directories and files (movies) within it
(top level only, there's no recursion).  Individual directories may be
counted or not; directories outside the default movie path must be
explicitly enabled for counting.  A directory that is not counted will
still show as "Directory"; a counted directory that has no movies and no
subdirectories will show as "Empty", or "No matches" if tags are being
filtered.