[config.py] Adjust ConfigClock to allow duration input
For some reason the ConfigClock code has been used as a base for duration period input.  When ConfigClock was updated for 12 and 24 hour clock operation this had an unintended impact on the Duration input screen.

This patch allows the Duration inputs to continue to use ConfigClock by forcing the input mode to 24 hour clock.  This should suffice until a cleaner implementation can be made.  This may happen as part of the forthcoming Setup.py refactor.