[MenuList.py] Enhance list movement methods
This change renames the recently added first and last methods to moveTop and moveBottom respectively. This better aligns with the underlying C++ code. The underlying C++ code is a little odd in that the functions provided moveTop and moveEnd are not natural opposites (which is why I originally chose the different names). I have gone with moveTop and moveBottom rather than moveStart and moveEnd as this works better with the other navigation methods.

Add new methods moveUp and moveDown as synonyms for the existing up and down methods. These additional new names also better match the underlying C++ code and offer symmetry with the equivalent methods in the ConfigList.py code.