Fix Bug #580: ChannelSelection context menu has active shortcuts for menu items that aren't available
Create new _append_when_current_valid() function that creates entries in an actions
map for the shortcut buttons to ensure that shortcut buttons are in
the self["actions"] ActionMap only if they have a corresponding menu entry.

Modify append_when_current_valid() to call _append_when_current_valid()
with actions set to None to emulate original behaviour of the function.
Print a deprecation warning.

Use the new _append_when_current_valid() function to ensure that
all button shortcuts correspond to a valid menu item. Only
unconditionally set shortcut buttons for unconditional menu items.

Change the creation of the "Add marker" menu item from an explicit
test on the validity of current to use _append_when_current_valid(),
since the test is equivalent in the function and it handles the
shortcut on "Add marker".