[MovieSelection] Display "friendly" locations
Replace the default movie location (typically "/media/hdd/movie") with
"Home".  Adjust the order of buildMovieLocationList to match the order
of do_move, as well as passing in an optional base directory to find
and list subdirectories.  This means do_move can use selectMovieLocation
and do_copy can use subdirectories.