Fix Bug #586: Horizontal menus in DreamPlex stuck
[Sources.List, Converter.StringList]

Pass through the wrap_around property from Renderer.Listbox.

This allows wrap_around to be overridden in the renderer from the source
and to have its value tested in the chain.

[Sources.List]

Allow the enableWrapAround argument to __init__() to be passed through to
the renderer so that it over-rides the renderer default and skin setting.

If the renderer's height is too small to allow the operations
moveUp/moveDown/pageUp/pageDown to operate, detect this and
fall back to the old code's (prior to the fix of bugs #575 & #576)
method, using the index property passed down to the renderer, and
using the previous paging distance guesses of +/-10 lines.